+++ src/main/java/dev/enjarai/trickster/item/ChannelItem.java
@@ -1,5 +1,7 @@
 package dev.enjarai.trickster.item;
 
+import java.util.Optional;
+
 import dev.enjarai.trickster.spell.EvaluationResult;
 import dev.enjarai.trickster.spell.Fragment;
 import dev.enjarai.trickster.spell.SpellContext;
@@ -7,7 +9,7 @@ import dev.enjarai.trickster.spell.trick.Trick;
 import net.minecraft.item.ItemStack;
 
 public interface ChannelItem {
-    EvaluationResult messageListenBehavior(Trick<?> trickSource, SpellContext ctx, ItemStack stack, int timeout);
+    EvaluationResult messageListenBehavior(Trick<?> trickSource, SpellContext ctx, ItemStack stack, Optional<Integer> timeout);
 
     void messageSendBehavior(Trick<?> trickSource, SpellContext ctx, ItemStack stack, Fragment value);