+++ src/main/java/dev/enjarai/trickster/spell/trick/misc/MessageListenTrick.java
@@ -20,17 +20,17 @@ import java.util.Optional;
 
 public class MessageListenTrick extends Trick<MessageListenTrick> {
     public MessageListenTrick() {
-        super(Pattern.of(4, 0, 7, 2, 4), Signature.of(FragmentType.NUMBER, MessageListenTrick::run, RetType.ANY.listOfRet().executor()));
-        overload(Signature.of(FragmentType.NUMBER, FragmentType.SLOT, MessageListenTrick::runWithChannel,
+        super(Pattern.of(4, 0, 7, 2, 4), Signature.of(FragmentType.NUMBER.optionalOfArg(), MessageListenTrick::run, RetType.ANY.listOfRet().executor()));
+        overload(Signature.of(FragmentType.NUMBER.optionalOfArg(), FragmentType.SLOT, MessageListenTrick::runWithChannel,
                 RetType.ANY.listOfRet().thisFunctionExistsSolelyForMessageListeningOnItemsBecauseWeAlreadyHadAnAbstractionForItAndWeReallyDontWantToReworkItSoThisWillHaveToDoHonestly()));
     }
 
     //TODO: how should we stop this from running in single-tick mode
-    public SpellExecutor run(SpellContext ctx, NumberFragment timeout) throws BlunderException {
-        return new MessageListenerSpellExecutor(ctx.state(), timeout.asInt(), Optional.empty());
+    public SpellExecutor run(SpellContext ctx, Optional<NumberFragment> timeout) throws BlunderException {
+        return new MessageListenerSpellExecutor(ctx.state(), timeout.map(n -> n.asInt()), Optional.empty());
     }
 
-    public EvaluationResult runWithChannel(SpellContext ctx, NumberFragment timeout, SlotFragment slot) throws BlunderException {
+    public EvaluationResult runWithChannel(SpellContext ctx, Optional<NumberFragment> timeout, SlotFragment slot) throws BlunderException {
         var stack = slot.reference(this, ctx);
         var range = ctx.source().getPos().distance(slot.getSourceOrCasterPos(this, ctx));
         var item = stack.getItem();
@@ -40,7 +40,7 @@ public class MessageListenTrick extends Trick<MessageListenTrick> {
                 throw new OutOfRangeBlunder(this, channelItem.getRange(), range);
             }
 
-            return channelItem.messageListenBehavior(this, ctx, stack, timeout.asInt());
+            return channelItem.messageListenBehavior(this, ctx, stack, timeout.map(n -> n.asInt()));
         }
 
         throw new ItemInvalidBlunder(this);