+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Made the string displayed while scrolling through the top hat translatable. (@Awakened-Redstone)
 - Added a new spell core which only runs while a player is near. (@enjarai and @aurakle)
 - Fixed a dumb crash when the minecraft window is resized to 0x0 pixels. (@enjarai)
+- Made the first argument of the message listening trick optional. (@aurakle)